Upgrade http libs, and reorganize.
