+++ manager/src/main/scala/org/broadinstitute/transporter/TransporterManager.scala
@@ -6,7 +6,7 @@ import doobie.util.ExecutionContexts
 import org.broadinstitute.transporter.web.{ManagerApi, SwaggerMiddleware}
 import org.broadinstitute.transporter.db.DbClient
 import org.broadinstitute.transporter.kafka.KafkaClient
-import org.broadinstitute.transporter.status.StatusController
+import org.broadinstitute.transporter.info.InfoController
 import org.http4s.implicits._
 import org.http4s.rho.swagger.models.Info
 import org.http4s.server.blaze.BlazeServerBuilder
@@ -16,6 +16,12 @@ import pureconfig.module.catseffect._
 import scala.concurrent.ExecutionContext
 
 object TransporterManager extends IOApp {
+
+  private val appInfo = Info(
+    title = "Transporter API",
+    version = BuildInfo.version,
+    description = Some("Bulk file-transfer system for data ingest / delivery")
+  )
 
   override def run(args: List[String]): IO[ExitCode] =
     loadConfigF[IO, ManagerConfig]("org.broadinstitute.transporter")
@@ -34,21 +40,16 @@ object TransporterManager extends IOApp {
     (dbResource, kafkaResource).tupled.use {
       case (dbClient, kafkaClient) =>
         val app = ManagerApp(
-          statusController = new StatusController(dbClient, kafkaClient)
+          infoController = new InfoController(appInfo.version, dbClient, kafkaClient)
         )
 
-        val appApi = new ManagerApi(BuildInfo.version, app)
-        val appInfo = Info(
-          title = "Transporter API",
-          version = BuildInfo.version,
-          description = Some("Bulk file-transfer system for data ingest / delivery")
-        )
+        val appApi = new ManagerApi(app)
 
         val routes = SwaggerMiddleware(appApi, appInfo, blockingEc).orNotFound
 
         BlazeServerBuilder[IO]
           .bindHttp(port = config.port, host = "0.0.0.0")
-          .withHttpApp(Logger(logHeaders = true, logBody = true)(routes))
+          .withHttpApp(Logger.httpApp(logHeaders = true, logBody = true)(routes))
           .serve
           .compile
           .lastOrError
