+++ manager/src/main/scala/org/broadinstitute/transporter/info/ManagerVersion.scala
@@ -1,0 +1,9 @@
+package org.broadinstitute.transporter.info
+
+import io.circe.Encoder
+
+case class ManagerVersion(version: String)
+
+object ManagerVersion {
+  implicit val encoder: Encoder[ManagerVersion] = io.circe.derivation.deriveEncoder
+}
