+++ manager/src/main/scala/org/broadinstitute/transporter/web/ManagerApi.scala
@@ -2,20 +2,14 @@ package org.broadinstitute.transporter.web
 
 import cats.effect.IO
 import org.broadinstitute.transporter.ManagerApp
-import org.broadinstitute.transporter.status.ManagerStatus
+import org.broadinstitute.transporter.info.{ManagerStatus, ManagerVersion}
 import org.http4s._
 import org.http4s.rho.RhoRoutes
 import org.http4s.rho.swagger.{syntax => swaggerSyntax}
 
-class ManagerApi(appVersion: String, app: ManagerApp) extends RhoRoutes[IO] {
+class ManagerApi(app: ManagerApp) extends RhoRoutes[IO] {
 
   import swaggerSyntax.io._
-
-  // This is in-sourced from `CirceEntityEncoder`, with pinned type params.
-  // It _should_ work to instead extend than trait, but for some reason that
-  // gives diverging implicits errors.
-  private implicit val statusEncoder: EntityEncoder[IO, ManagerStatus] =
-    org.http4s.circe.jsonEncoderOf
 
   private val statusRoute =
     "Query operational status of the system" ** GET / "status"
@@ -23,9 +17,26 @@ class ManagerApi(appVersion: String, app: ManagerApp) extends RhoRoutes[IO] {
   private val versionRoute =
     "Query version of the system" ** GET / "version"
 
-  // NOTE: Has to be a thunk here to prevent eager evaluation / caching.
+  /*
+   * NOTES on weirdness discovered in Rho that you'll see below:
+   *
+   *   1. Thunks are required on no-arg routes to prevent eager
+   *      evaluation & result caching on app startup. This is
+   *      desired behavior according to the Rho maintainers.
+   *
+   *   2. http4s provides a generic method for deriving `EntityEncoder`
+   *      instances for any type with a circe `Encoder`, but for some
+   *      reason combining it w/ the thunks causes divergence in
+   *      implicit search. There are a couple maybe-related bugs filed
+   *      in Rho; I'll also try to minimize & file a bug. Until then,
+   *      we have to live with the boilerplate.
+   */
+
   statusRoute |>> { () =>
-    app.statusController.status.flatMap { status =>
+    implicit val encoder: EntityEncoder[IO, ManagerStatus] =
+      org.http4s.circe.jsonEncoderOf
+
+    app.infoController.status.flatMap { status =>
       if (status.ok) {
         Ok(status)
       } else {
@@ -34,7 +45,10 @@ class ManagerApi(appVersion: String, app: ManagerApp) extends RhoRoutes[IO] {
     }
   }
 
-  // NOTE: The response here is evaluated at startup and cached,
-  // since it's not a thunk.
-  versionRoute |>> Ok(appVersion)
+  versionRoute |>> { () =>
+    implicit val encoder: EntityEncoder[IO, ManagerVersion] =
+      org.http4s.circe.jsonEncoderOf
+
+    app.infoController.version.flatMap(Ok(_))
+  }
 }
